Module c19::agent::default

Default implementation of the Agent trait.

This is a simple yet decent implementation of the Agent trait and should answer most use cases.

This agent exposes a GET and a PUT endpoints to allow an app to get and set values from and to the state.

The format of the keys are dependent on the State layer being used. For example, when using the Default state, the keys are expected to be a String and the values are expected to be in the format specified in the documentation of the state layer.

GET /<key>

To get a value, the app can send a GET request with a key that represents a key in the state.

Expects key to be a String. Returns the value as-is from the state.

Example

Assuming usage of the Default state, a value might look like this:

GET /cat

{"ts":1601241450390,"ttl":null,"value":"garfield"}

As you can see, the value holds more than the value itself. Included are the ttl (can be null) which is an absolute value of when this key will expire, and the timestamp (ts) that this key was created.

A different form of the value might be returned, depending on which state layer is being used. In any case, this agent implementation does not assume anything about the format of the values returned by the state.

PUT /

To set a value to the state, the app can send a PUT request with a body that conforms to the state expected value.

The agent does not assume anything about the format of the value in the body of the message. The value is passed to the state as-is and it is up to the app to make sure it conforms to the expected format by the state.

For example, using the Default state, an app would send a PUT request like so:

A JSON body of the followibng format:

{"cat": {"value": "garfield", "ttl": 60000}}

According to the Default state documentation, the value can be anything JSON. ttl is optional and another field ts can be specified to override the automatic ts set by the state.

Please refer to the Default state implementation for more information.

Structs

Default
The Default struct.

Statics

__init215733339167654759

Functions

get_handler
Returns the value associated with the given key.

handler
Accepts a request and dynamically dispatches the handler based on the method of the request.

set_handler
Sets the key and value specified in the request.

Type Definitions

Result