Function c19::agent::default::get_handler

fn get_handler(state: Arc<dyn State>, req: &Request<Body>) -> Response<Body>

Returns the value associated with the given key.

GET /<key>

Expects key to be a String. Returns the value as-is from the state.

Example

Assuming usage of the Default state, a value might look like this:

GET /cat

{"ts":1601241450390,"ttl":null,"value":"garfield"}

As you can see, the value holds more than the value itself. Included are the ttl (can be null) which is an absolute value of when this key will expire, and the timestamp (ts) that this key was created.

A different form of the value might be returned, depending on which state layer is being used. In any case, this agent implementation does not assume anything about the format of the values returned by the state.