Function c19::agent::default::handler

async fn handler(
    state: Arc<dyn State>, 
    req: Request<Body>
) -> Result<Response<Body>, Box<dyn StdError + Send + Sync>>

Accepts a request and dynamically dispatches the handler based on the method of the request.

Returns whatever the get and set handlers return or 404 (not found) if method is invalid.