Function c19::agent::default::set_handler

fn set_handler(
    state: Arc<dyn State>, 
    req: Request<Body>
) -> impl FutureExt<Output = Result<Response<Body>, Box<dyn StdError + Send + Sync>>>

Sets the key and value specified in the request.

PUT /

The agent does not assume anything about the format of the value in the body of the message. The value is passed to the state as-is and it is up to the app to make sure it conforms to the expected format by the state.

For example, using the Default state, an app would send a PUT request like so:

A JSON body of the followibng format:

{"cat": {"value": "garfield", "ttl": 60000}}

According to the Default state documentation, the value can be anything JSON. ttl is optional and another field ts can be specified to override the automatic ts set by the state.

Please refer to the Default state implementation for more information.