Module c19::config

The C19 configuration.

The configuration automatically initializes and holds the instances of all layers (Agent, State and Connection).

A YAML formatted configuration file is automatically loaded when the c19 process starts. The path to the configuration file is specified using the --config flag when running the process.

Examples:

The configuration it based on the different layers chosen. Evey layer implementation has its own configuration.

Here's a small example for a configuration file:

version: 0.1
spec:
  agent:
    kind: Default
    port: 3097
  state:
    kind: Default
    ttl: null
    purge_interval: 10000
    data_seeder:
      kind: File
      filename: data.json
  connection:
    kind: Default
    push_interval: 1000
    pull_interval: 60000
    force_publish: 0.1
    port: 4097
    target_port: 4098
    r0: 6
    timeout: 5000
    peer_provider:
      kind: Static
      peers:
        - 127.0.0.1

The configuration file is devided into three parts, one for each layer. The configuration for each layer is specific to that layer based on its kind. In this example you can see that the configuration will load the Default agent, Default state and Default connection and within the connection layer it will load the static peer provider.

See the documentation for each layer implementation to find out what available configuration settings there are.

Structs

Config

Spec

Constants

DEFAULT_CONFIG_FILE

Functions

new
Returns the configuration after dynanmically loading all layers.