Module c19::connection

The Connection layer.

The connection layer is responsible for exchaging the state with other peers.

There can be many implementations for the connection layer. One could choose the protocol, the transport layer, how to choose peers to exchange the state with, the rate in which the state is being exchanged, compression, etc...

See the default connection layer implementation for some ideas.

Modules

default
The default connection layer implementation

peer_provider
Peer provider.

Constants

_Connection_registry

Traits

Connection
The Connection Trait.