Module c19::connection::default

The default connection layer implementation

This is an implementation of the Connection trait.

It offers a simple and powerful way of exchanging data with other peers.

Choosing peers

The connection chooses peers by using a peer provider. A peer provider is an abstration over a vector of peers. A peer provider is responsible for retrieving the full list of peers this connection layer can connect with. See the peer provider documentation for more information.

If not provided in the configuration, this connection will choose the k8s peer provider to get the full list of available peers to connect to. The k8s peer provider queries Kubernetes server api for endpoints of other peers by using the selector field in the configuration.

The default connection will then randomly select a subset of the peers and will exchange the state with each one of them in parallel. The number of peers sampled from the full list is determined by the r0 configuration field.

Exchanging data

The connection uses HTTP to push and pull (PUT and GET) the state to and from other peers.

Pushing

Only the changes from the last publish time will be pushed to other peers.

Pulling

When pulling, the connection layer specifies its own state version and if it matches the one that the other peer has then nothing will be exchanged. If the versions do not match then the other peer will return its own full state.

The connection layer does not assume anything about the content of the data being exchanged. The data will be passed as-in to other peers.

The interval in which the data will be exchanged is set in the push_interval and pull_interval configuration flags.

See more about the default implementation and the different options it provides in the struct documentation.

Structs

Default

Statics

__init8791889224756635807

Functions

get_handler
Returns an HTTP response with the full state as JSON.

handler

set_handler
Accepts a JSON body that represents a state. Merges it with its own state.

Type Definitions

Result