Struct c19::connection::default::Default

pub struct Default {
    port: u16,
    target_port: Option<u16>,
    push_interval: u64,
    pull_interval: u64,
    r0: usize,
    timeout: u64,
    pub peer_provider: Box<dyn PeerProvider>,
}

Fields
port: u16

The port to bind to and listen for connection from other peers. Default port: 4097.

target_port: Option<u16>

An optional port to use as a target when sending the state to peers. If ommited, then the port field will be used both for accepting connections and when connecting to other peers. Default is null (use the port field).

push_interval: u64

The publish interval in milliseconds. Default value is 1 second (1000 milliseconds).

The connection layer will publish a new state push_interval after it has finished publishing the previous state. This means that if preparing a state for publishing takes a few seconds, then only after those few seconds the counter will start counting push_interval.

pull_interval: u64

The pull interval in milliseconds. Default value is 60 seconds (60000 milliseconds).

The connection layer will connect and pull the state from peers every pull_interval.

r0: usize

The number of peers to connect to on each interval and exchange the state with. Default value: 3.

timeout: u64

The connection timeout when connecting and exchanging data with a peer. Default value: 1000ms.

peer_provider: Box<dyn PeerProvider>

The peer provider to use.

The connection layer will reach out to the peer provider to get the full list of available peers to choose from.

Default value: k8s.

Implementations

impl Default

pub(in connection::default) async fn server<'_>(
&'_ self,
state: Arc<dyn State>
) -> Result<(), Box<dyn StdError + Send + Sync>>

pub(in connection::default) async fn receiver<'_>(
&'_ self,
state: Arc<dyn State>
) -> Result<(), Box<dyn StdError + Send + Sync>>

Receiver thread. Connects to r0 peers at random and pulls their state into its own state.

When pulling the state, the version of the current state is specified in the request path so the other peer would determine whether to respond with its full state or nothing.

If the current version matches the one of the peer's version then the peer will respond with 204 (no content). The full state will be returned by the peer if the versions do not match.

pub(in connection::default) async fn publisher<'_>(
&'_ self,
state: Arc<dyn State>
) -> Result<(), Box<dyn StdError + Send + Sync>>

Publisher thread. Publishes the state at an interval time.

This thread will loop forever, waiting for interval time to pass. It will then reach out to the peer provider to get the full list of available peers and will randomly pick r0 peers to exchange the state with.

It connects to other peers in parallel and publishes the changes since last publish.