Module c19::connection::peer_provider

Peer provider.

The peer provider is responsible for retreiving peers that are available for exchanging states.

The peer provider is specific to the default connection implementation, although it can serve other implemntations in the future. It is implemented as a trait to allow the default connection implementation to use different kind of providers, not only the K8s, which is the default one.

The provider is first loaded by the configuration and then initialized by the default connection itself. It is then used by the default connection to get the list of available peers to choose from.

Modules

k8s
An implementation of the peer provider.

static_peer_provider
An implementation of the peer provider.

Enums

Peer

Constants

_PeerProvider_registry

Traits

PeerProvider