Module c19::connection::peer_provider::k8s

An implementation of the peer provider.

This implementation targets Kubernetes. It queries the Kubernetes api server for endpoints of pods based on the selector configuration.

Watches the list of pods for any change and updates a local vector of peers so when the conncetion implementation queries the peer provider, it returns the full vector.

Example:

   peer_provider:
     kind: K8s
     selector:
       c19: example
     namespace: :all

Structs

K8s

Statics

__init12866312320927563797

Type Definitions

Result