Trait c19::connection::Connection

pub trait Connection: Debug + Send + Sync + Serialize + Deserialize {
    pub fn start<'a>(
        &'a self, 
        state: Arc<dyn State>
    ) -> BoxFuture<'a, Result<(), Box<dyn StdError + Send + Sync>>>;
}

The Connection Trait.

The only required method is the start method where the state is passed to the implementation to should be used to set and get values to and from the state.

See the default connection implementation for an idea on how to implement a connection layer.

Required methods

pub fn start<'a>(
&'a self,
state: Arc<dyn State>
) -> BoxFuture<'a, Result<(), Box<dyn StdError + Send + Sync>>>

Loading content...