Module c19::state

The State layer.

The state is responsible for holding the data and for exposing a way to get and set values to and from it.

The State trait does its best to assume nothing about the data an implementor might hold. It can be a Key/Value store like the default State implementation, a blob of binary data or anything else an implementor wishes for.

It does so by using a StateValue trait with default implementations of serializing the data to different formats. For example, the default state implementation implements a serialization of StateValue to JSON.

The state layer can choose to implement other mechanism related to the state, like TTL, data compression, etc.

Modules

data_seeder

default
The default implementation of the State layer.

Constants

_State_registry

Traits

CloneState

State
The State trait.

StateValue
A trait used as a protocol between different components that use the state.

Type Definitions

SafeState
An atomic reference to a state.