Module c19::state::default

The default implementation of the State layer.

This implementation is of a Key/Value state where a value can be anything that can be serialized into a JSON representation.

The default state has the following behavior:

It's a Key/Value store

It supports TTL

It resolves conflicts using a timestamp

It records version history

Keys

This state expects keys to be of a string type.

Values

The values are expected to be a JSON of the following form:

{"value": <anything JSON>, "ttl": <optional ttl in milliseconds>, "ts": <optional, manual
setting of the timestamp of the value>}

value

The only required field is the value which can be anything JSON. Even a JSON object.

ttl

The ttl field is optional. If not used then the default ttl from the state configuration will be used, if one specified.

ts

The ts field is optional and can be used to override the timestamp that is automatically set for every new value.

TTL

The state returns None for expired keys. Expired keys are filtered out when the storage is iterated and are purged by a thread running in the background.

The purger thread uses the interval settings from the configuration of the state.

Conflicts

Since this is a distributed system, the state might be updated by different peers that are not yet in sync. To resolve a conflict where a key is being updated by more than one peer, a timestamp is used. Only a newer key can override an older one. The timestamp should be the timestamp when the key was first created (by the source).

Version History

The state records version history for every change that is made to the state. To make sure the version history doesn't get bloated it is being purged on every version_ttl milliseconds.

See the Default state struct for details on the different fields and configurations.

Structs

Default
The default state struct.

Value
An implementation of the StateValue trait.

Version
Version information.

Constants

MAX_SET_OPS
The maximum number of pending async set operations.

Statics

__init5504723193248535467

Functions

async_set
Async set thread.

hash

purge
Purges expired keys at a specficied interval.