Struct c19::state::default::Default

pub struct Default {
    ttl: Option<u64>,
    purge_interval: u64,
    data_seeder: Option<Arc<RwLock<Box<dyn DataSeeder>>>>,
    version: Arc<RwLock<String>>,
    tx: Option<SyncSender<Vec<u8>>>,
    storage: Arc<RwLock<HashMap<String, Box<Value>>>>,
    is_dirty: Arc<RwLock<bool>>,
}

The default state struct.

Fields
ttl: Option<u64>

The default TTL (in milliseconds) to use if none is specified when setting a new value.

purge_interval: u64

The interval in milliseconds in which to purge expired values.

Default value is 1 minute (60000 milliseconds).

data_seeder: Option<Arc<RwLock<Box<dyn DataSeeder>>>>

The DataSeeder to use for seeding the data on initialization.

version: Arc<RwLock<String>>

The version of the current state.

This is set to a random unique string on every state change.

tx: Option<SyncSender<Vec<u8>>>

The SyncSender channel to use for async set operations

When a set operation is being commited to the state, the state will pass the operation to an async handler which will then commit the changes to the state.

storage: Arc<RwLock<HashMap<String, Box<Value>>>>

The data storage in the form of a Key/Value hashmap.

is_dirty: Arc<RwLock<bool>>

Calculating the version is a bit expensive so we use the dirty flag to lazily calculate the verison on-demand.

Implementations

impl Default

fn set(&self, map: &HashMap<String, Box<Value>>)

Merges the two maps while resolving conflicts.

A value from the other map will be commited to the state only if it has a newer timestamp.

If there was a change to the sate, the version will be recorded in the version history.

fn purge(&self)

Purges expired keys.

fn seed(
&self,
data_seeder: Arc<RwLock<Box<dyn DataSeeder>>>
) -> Result<(), Box<dyn StdError>>

Seeds the state with the data from the DataSeeder.