Trait c19::state::State

pub trait State: Debug + Send + Sync + CloneState + Serialize + Deserialize {
    pub fn init(&self) -> Arc<dyn State>;

    pub fn version(&self) -> String;

    pub fn set(&self, value: &dyn StateValue) -> Result<(), Box<dyn StdError>>;

    pub fn get(&self, key: &dyn StateValue) -> Option<Box<dyn StateValue>>;

    pub fn diff(
        &self, 
        other: &dyn StateValue
    ) -> Result<Box<dyn StateValue>, Box<dyn StdError>>;

    pub fn get_root(&self) -> Option<Box<dyn StateValue>>;

    pub fn get_or(
        &self, 
        key: &dyn StateValue, 
        default: Box<dyn StateValue>
    ) -> Box<dyn StateValue> { ... }
}

The State trait.

Every state implementor must implement this trait. It is first auto-loaded by the configuration deserializer and then initialized by the library by calling the init function. The init function should return a SafeState which is then passed to the connection and agent layers.

Required methods

pub fn init(&self) -> Arc<dyn State>

Initializes the state and returns a state that is safe to be shared across threads.

A state object is already loaded by the configuration. The implementor can use this function to add or initialize any other relevant data and then return a SafeState object which is shared with the connection and agent layers.

pub fn version(&self) -> String

Returns the version of the current state.

An implementor can use this function to keep a version for each "state" of the sate. For example, the default state implementation sets this value to a random string whenever the state changes. It is then saves that version in a version history which allows for the connection layer to compute diff between two versions.

pub fn set(&self, value: &dyn StateValue) -> Result<(), Box<dyn StdError>>

Sets a value to the state.

There's no assumption about the value itself. It can be anything the implementor wishes. The default state implementation, for example, treats this value as a map of key/value pairs where the key is a String and the value conforms to a serde_json::Value value.

pub fn get(&self, key: &dyn StateValue) -> Option<Box<dyn StateValue>>

Gets the value associated with the specified key.

To allow maximum flexibility, the key itself is a StateValue, which in effect means it can be anything desired by the implementor.

pub fn diff(
&self,
other: &dyn StateValue
) -> Result<Box<dyn StateValue>, Box<dyn StdError>>

Returns the difference between this and the other state.

pub fn get_root(&self) -> Option<Box<dyn StateValue>>

Returns the whole state as a StateValue.

This is helpful when the connection layer wishes to publish the whole state to its peers.

Loading content...
Provided methods

pub fn get_or(
&self,
key: &dyn StateValue,
default: Box<dyn StateValue>
) -> Box<dyn StateValue>

Returns the value associated with the specified key or the default if the key was not found in the state.

Loading content...