Trait c19::state::StateValue

pub trait StateValue: Send + Sync {
    pub fn as_bytes(&self) -> Option<Vec<u8>>;
}

A trait used as a protocol between different components that use the state.

Both the Connection and Agent layers use the state to get and set values. They do not know anything about the implementation of one another. This trait allows them to use the same state without assuming anything about the structure of a state value.

The trait offers functions to serialize to different formats. The default connection layer for example, serializes a state value to json to be exchanged with other peers.

The functions in this trait have a default implementation that always return None so that implementors don't have to bother and implement all serialization functions that are irrelevant to them. If an implementor of an Agent layer, for example, tries to serialize a state value that the State layer does not support, it will get a None value in response. This might indicate an incompatible usage of an Agent and State layers.

When a user of the c19 protocol chooses their state, agent and connection layers, they should ensure that each component is compatible with one another.

An implementor of layer might choose to run different compatability tests at startup and notify the user of incompatabilities.

Required methods

pub fn as_bytes(&self) -> Option<Vec<u8>>

Loading content...